Working Serialize
Changed the Ids to be u32 (json doesn't work nice with u64).
Manually made the CRC32 for each component id as an example, inlcuding
CRC32 is outside the scope.
Added Manager::Insert via Json structure.

+++ src/component_types.h
@@ -4,8 +4,8 @@ namespace minmod
 {
     namespace component
     {
-        using OwnerId = std::size_t;
-        using Id = std::size_t;
+        using OwnerId = std::int32_t;
+        using Id = std::uint32_t;
         static const Id INVALID_ID = 0;
     }
 }
